# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SyllabusAI API",
        "description": "AI powered learning",
        "version": "1.0.0"
    },
    "paths": {
        "/user/auth": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Auth",
                "operationId": "auth",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Me",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/generation/chapters": {
            "post": {
                "tags": [
                    "LLM Generation"
                ],
                "summary": "Generate Chapters",
                "operationId": "generate_chapters",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateChaptersRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChaptersGenerationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/generation/chapter": {
            "post": {
                "tags": [
                    "LLM Generation"
                ],
                "summary": "Generate Chapter",
                "operationId": "generate_chapter",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateChapterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenerateChapterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/generation/quiz": {
            "post": {
                "tags": [
                    "LLM Generation"
                ],
                "summary": "Generate Quiz",
                "operationId": "generate_quiz",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateQuizRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenerateQuizResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/internal/health": {
            "get": {
                "tags": [
                    "Internal"
                ],
                "summary": "Health",
                "operationId": "health",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/internal/cache": {
            "post": {
                "tags": [
                    "Internal"
                ],
                "summary": "Invalidate Cache",
                "operationId": "invalidate_cache",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AuthRequest": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "AuthRequest"
            },
            "ChaptersGenerationResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "ChaptersGenerationResponse"
            },
            "GenerateChapterRequest": {
                "properties": {
                    "chapter": {
                        "type": "string",
                        "title": "Chapter"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "chapter",
                    "description"
                ],
                "title": "GenerateChapterRequest"
            },
            "GenerateChapterResponse": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "GenerateChapterResponse"
            },
            "GenerateChaptersRequest": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "topic": {
                        "type": "string",
                        "title": "Topic"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "topic"
                ],
                "title": "GenerateChaptersRequest"
            },
            "GenerateQuizRequest": {
                "properties": {
                    "topic": {
                        "type": "string",
                        "title": "Topic"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "topic",
                    "content"
                ],
                "title": "GenerateQuizRequest"
            },
            "GenerateQuizResponse": {
                "properties": {
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        },
                        "type": "array",
                        "title": "Questions"
                    }
                },
                "type": "object",
                "required": [
                    "questions"
                ],
                "title": "GenerateQuizResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Question": {
                "properties": {
                    "question": {
                        "type": "string",
                        "title": "Question"
                    },
                    "correctOptionIndex": {
                        "type": "integer",
                        "title": "Correctoptionindex"
                    },
                    "options": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Options"
                    }
                },
                "type": "object",
                "required": [
                    "question",
                    "correctOptionIndex",
                    "options"
                ],
                "title": "Question"
            },
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token"
                    },
                    "picture": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Picture"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "name",
                    "is_active"
                ],
                "title": "UserResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}